pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "eheimdigital"
version = "1.0.0"
dependencies = ["websocket-client"]
